+++ Arkanoid/Arkanoid/Grid.hpp
@@ -26,7 +26,7 @@ struct Grid {
         for (int r = 0; r < num_row ; r++) {
             for (int c = 0; c < num_col; c++) {
                 Brick& b = get_brick(c, r);
-                b.set_pos(c * g_brick_width + g_offset_x, r * g_brick_height + g_offset_y);
+                b.set_pos((int) c * g_brick_width + g_offset_x, (int) r * g_brick_height + g_offset_y);
                 
                 b.hp = (num_row - r) % b.max_hp;
                 if (!b.hp) b.hp = 1;
